+++ src/mol-repr/structure/visual/bond-intra-unit-cylinder.ts
@@ -149,10 +149,8 @@ function getIntraUnitBondCylinderBuilderProps(unit: Unit.Atomic, structure: Stru
                 if (isBondType(f, BondType.Flag.Aromatic) || (arCount && !ignoreComputedAromatic)) {
                     if (arCount === 2) {
                         return LinkStyle.MirroredAromatic;
-                    } else if (arCount === 1 || deloTriplets?.getThirdElement(aI, bI)) {
-                        return LinkStyle.Aromatic;
                     } else {
-                        // case for bonds between two aromatic rings
+                        return LinkStyle.Aromatic;
                     }
                 }
             }